refactor(cli): import command modules directly rather than through barrels
Five handlers reached their orchestrators through `../registry`,
`../data-transfer` and `./index`. The direct path names the module that actually
owns the function, which is the clearer spelling and stops a handler pulling in
every sibling the barrel re-exports.

The two tests that stubbed those orchestrators mocked the barrel, so they
silently stopped intercepting once the handler no longer went through it — both
retargeted at the module the handler imports.
