+++ packages/cli/src/commands/add/handler.ts
@@ -11,8 +11,8 @@ import { reportLintIgnoreOutcomes } from "../../util/lint-ignore-report";
 import type { Logger } from "../../util/logger";
 import type { TextPrompt } from "../../util/tui-prompts";
 import { tuiSelect, tuiText } from "../../util/tui-prompts";
-import { runAddCommand } from "../registry";
 import { isCustomRegistrySource } from "../registry/apply";
+import { runAddCommand } from "../registry/commands";
 import type { RegistryManifest } from "../registry/types";
 import { deriveDatabaseName, promptDatabaseName, sanitizeDatabaseName, withAuthDatabaseName } from "./auth-database";
 import type { FeatureItem, NormalizedFeature } from "./features";
