+++ packages/cli/src/commands/import/handler.ts
@@ -2,8 +2,8 @@ import type { CommandHandler } from "../../util/command";
 import { defineHandler } from "../../util/command";
 import { EXIT_CODE } from "../../util/exit-code";
 import { resolveProductionWorkerUrl } from "../../util/resolve-target";
-import type { ImportCommandData } from "../data-transfer";
-import { runImportCommand } from "../data-transfer";
+import type { ImportCommandData } from "../data-transfer/import";
+import { runImportCommand } from "../data-transfer/import";
 import type { ImportSourceName } from "../data-transfer/import-source";
 import { IMPORT_SOURCE_NAMES } from "../data-transfer/import-source";
 import type { ImportOptions } from "./index";
